improve convention names
