+++ src/blueprints/Do.Blueprints.Service.Application/Business/Default/RestApiConventions/AddChildToChildrenConvention.cs
@@ -1,0 +1,14 @@
+﻿using Do.RestApi.Configuration;
+using Humanizer;
+
+namespace Do.Business.Default.RestApiConventions;
+
+public class AddChildToChildrenConvention : IApiModelConvention<ActionModelContext>
+{
+    public void Apply(ActionModelContext context)
+    {
+        if (!context.Action.Id.StartsWith("Add")) { return; }
+
+        context.Action.Route = context.Action.Route.Replace(context.Action.Id, context.Action.Id[3..].Pluralize());
+    }
+}
